frontend/requirements.txt
streamlit
requests
python-dotenv
python-docx
pypdf2
pandas
